requirements.txt
xrpl-py==4.6.0b0
python-dotenv==1.2.1
requests==2.32.5
Flask==3.1.3